dx_analysis/requirements.txt
flask==3.0.0
pdfminer.six==20221105
openai>=1.0.0
python-dotenv==1.0.0 